Failures were specifically seen in ansible installation in FIPs caused by python2-django from epel being installed instead of pulp.
Increase priority of pulp.repos to precede epel for ci installation
Adding a priorty of 50 (default undeclared is 99) to have pulp.repo
installation of RPMs prioritized over other repos during installation.

Update the jjb/jobs/macros.yaml to check if FIPS is enabled. If
enabled, install yum-plugin-priorities. This enables the priority
field in pulp.repo (set to 50) during pulp_server installation. The
pulp patched and built RPMs will be installed which have been modified
for FIPs.

For users manually running the pulp-fips installation as part of
pulp_server install, updated the handler to install
yum-plugin-prorities as part of the ansible playbook.

See: #4357